// Qanat.Web/src/app/shared/generated/model/support-ticket-status-update-dto.ts
/**
 * Qanat.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class SupportTicketStatusUpdateDto { 
    SupportTicketStatusID: number;
    SupportTicketPriorityID: number;
    AssignedUserID?: number | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface SupportTicketStatusUpdateDtoForm { 
    SupportTicketStatusID: FormControl<number>;
    SupportTicketPriorityID: FormControl<number>;
    AssignedUserID?: FormControl<number>;
}

export class SupportTicketStatusUpdateDtoFormControls { 
    public static SupportTicketStatusID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
            ],
        }
    );
    public static SupportTicketPriorityID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
            ],
        }
    );
    public static AssignedUserID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}